refactor: Move achievement calculation logic to StartPlayPresenter
